backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-multipart
sentence-transformers
faiss-cpu
openai
anthropic
python-dotenv
numpy
cryptography
neo4j
